+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## Unreleased
+
 ## 0.2.1
 
+* Improve health check endpoint to include status of various features
 * Change path handling for chat completions streams to put separate requests into separate streams. Requests can pass a UUID as the suffix of the path now, but is backwards compatible with an older version of the frontend code.
 
 ## 0.2.0